Split DbException and MatcherException into separate files
