+++ newsboat.cpp
@@ -9,8 +9,9 @@
 #include "cliargsparser.h"
 #include "config.h"
 #include "controller.h"
+#include "dbexception.h"
 #include "exception.h"
-#include "exceptions.h"
+#include "matcherexception.h"
 #include "rss.h"
 #include "rss/parser.h"
 #include "view.h"
