+++ src/cache.cpp
@@ -12,8 +12,9 @@
 #include "config.h"
 #include "configcontainer.h"
 #include "controller.h"
-#include "exceptions.h"
+#include "dbexception.h"
 #include "logger.h"
+#include "matcherexception.h"
 #include "rss.h"
 #include "strprintf.h"
 #include "utils.h"
