+++ src/colormanager.cpp
@@ -2,13 +2,13 @@
 
 #include "config.h"
 #include "confighandlerexception.h"
-#include "exceptions.h"
 #include "feedlistformaction.h"
 #include "filebrowserformaction.h"
 #include "helpformaction.h"
 #include "itemlistformaction.h"
 #include "itemviewformaction.h"
 #include "logger.h"
+#include "matcherexception.h"
 #include "pbview.h"
 #include "selectformaction.h"
 #include "strprintf.h"
