+++ src/controller.cpp
@@ -29,9 +29,9 @@
 #include "configcontainer.h"
 #include "configexception.h"
 #include "configparser.h"
+#include "dbexception.h"
 #include "downloadthread.h"
 #include "exception.h"
-#include "exceptions.h"
 #include "feedhqapi.h"
 #include "fileurlreader.h"
 #include "globals.h"
