+++ src/matcher.cpp
@@ -7,8 +7,8 @@
 #include <sys/time.h>
 #include <vector>
 
-#include "exceptions.h"
 #include "logger.h"
+#include "matcherexception.h"
 #include "utils.h"
 
 namespace newsboat {
