+++ src/pbcontroller.cpp
@@ -16,10 +16,10 @@
 #include "config.h"
 #include "configcontainer.h"
 #include "configexception.h"
-#include "exceptions.h"
 #include "globals.h"
 #include "keymap.h"
 #include "logger.h"
+#include "matcherexception.h"
 #include "nullconfigactionhandler.h"
 #include "pbview.h"
 #include "poddlthread.h"
