+++ src/view.cpp
@@ -24,15 +24,15 @@ extern "C" {
 }
 
 #include "config.h"
+#include "dbexception.h"
 #include "dialogs.h"
 #include "dialogsformaction.h"
 #include "exception.h"
-#include "exceptions.h"
 #include "feedlist.h"
 #include "feedlistformaction.h"
 #include "filebrowser.h"
-#include "formaction.h"
 #include "fmtstrformatter.h"
+#include "formaction.h"
 #include "help.h"
 #include "helpformaction.h"
 #include "htmlrenderer.h"
@@ -40,6 +40,7 @@ extern "C" {
 #include "itemlistformaction.h"
 #include "itemview.h"
 #include "itemviewformaction.h"
+#include "matcherexception.h"
 #include "keymap.h"
 #include "logger.h"
 #include "regexmanager.h"
